requirements.txt
# ============================================================
# Skill Barter API — Python Dependencies
# ============================================================
# Install:  pip install -r requirements.txt
# Python:   >= 3.11
# ============================================================

# ── Web Framework ─────────────────────────────────────────
fastapi==0.111.0          # ASGI web framework
uvicorn[standard]==0.29.0 # ASGI server (includes websockets + httptools)
python-multipart==0.0.9   # Required for FastAPI form data / file uploads

# ── Database ──────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30   # Async ORM (2.x has full async support)
asyncpg==0.29.0               # Async PostgreSQL driver (used by SQLAlchemy)
alembic==1.13.1               # DB migrations

# ── Validation ────────────────────────────────────────────
pydantic==2.7.1               # Data validation (v2 has 5-17x speedup vs v1)
pydantic-settings==2.2.1      # .env file loading via BaseSettings
email-validator==2.1.1        # Required for Pydantic EmailStr

# ── Authentication ────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT encode/decode (HS256 + RS256)
passlib[bcrypt]==1.7.4             # Password hashing — bcrypt backend
bcrypt==4.1.3                      # Explicit bcrypt pin (passlib backend)

# ── Cache & Message Broker ────────────────────────────────
redis[hiredis]==5.0.4         # Redis client (hiredis = C parser, 10x faster)
# Note: used for BOTH cache and Celery broker

# ── Background Jobs ───────────────────────────────────────
celery==5.4.0                 # Distributed task queue
flower==2.0.1                 # Celery monitoring UI (optional, remove in prod)

# ── Rate Limiting ─────────────────────────────────────────
slowapi==0.1.9                # Rate limiting for FastAPI (wraps limits library)
limits==3.12.0                # Storage backends for slowapi

# ── Logging ───────────────────────────────────────────────
structlog==24.1.0             # Structured JSON logging (plays well with cloud)

# ── HTTP Client (for internal service calls / health checks) ──
httpx==0.27.0                 # Async HTTP client

# ── CORS / Security Headers ───────────────────────────────
# Built into FastAPI via starlette — no extra package needed

# ── Dev / Test Dependencies ───────────────────────────────
pytest==8.2.0                 # Test runner
pytest-asyncio==0.23.6        # Async test support
httpx==0.27.0                 # Also used as async test client
factory-boy==3.3.0            # Test fixture factories
faker==25.0.0                 # Fake data generation for tests
pytest-cov==5.0.0             # Coverage reports

# ── Linting / Formatting ──────────────────────────────────
ruff==0.4.3                   # Ultra-fast linter + formatter (replaces flake8/isort/black)
mypy==1.10.0                  # Static type checking
